+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * New defcustom `orgbox-start-of-day` allows users to customize the
   start time of day.
+* New defcustom `orgbox-start-of-week` allows users to customize the
+  start day of week.
 * New defcustom `orgbox-start-of-weekends` allows users to customize
   the start time of weekends.
 * New defcustom `orgbox-start-of-evening` allows users to customize